fix move
